+++ R/get_game_data.R
@@ -30,6 +30,9 @@
 #' }
 get_4th_plays <- function(gid) {
 
+  # triggers download of games and fd model, if necessary
+  init_nfl4th()
+
   df <- .games_nfl4th %>%
     filter(game_id == gid)